refactor: Session, token interceptor의 authenticate 중복 제거
